+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Breaking changes
 
+- `x/escrow`: `CreateMsg.Source` is no longer optional and is now required.
 - `x/sigs`: `BumpSequenceMsg.User` is now required and cannot be empty.
 
 ## 0.24.1
@@ -28,6 +29,7 @@ Breaking changes
     new index implementation.
 
 Breaking changes
+
 - `orm.Index` is an interface now to allow multiple implementations. The new
   interface is a subset of the old structure, simplified and updated to support
   lazy loading: